+++ Nagisa.Shared/M2.Helpers.WinRT.cpp
@@ -10,24 +10,32 @@ License: The MIT License
 #include "M2.Helpers.WinRT.h"
 
 #include <Windows.h>
-#include <wrl/client.h>
+#include <wrl\client.h>
+#include <wrl\implements.h>
 
-#include <asyncinfo.h>
 #include <robuffer.h>
+#include <windows.foundation.h>
+#include <windows.storage.streams.h>
 
-using Microsoft::WRL::ComPtr;
-using ABI::Windows::Foundation::AsyncStatus;
-using Windows::Storage::Streams::IBuffer;
-using Windows::Storage::Streams::IBufferByteAccess;
+#include <string>
 
-// The m2_await_internal function uses the non-blocking way to try to wait 
+using ::Microsoft::WRL::MakeAndInitialize;
+using ::Microsoft::WRL::RuntimeClass;
+using ::Microsoft::WRL::RuntimeClassFlags;
+using ::Microsoft::WRL::RuntimeClassType;
+
+using abi_AsyncStatus = ::ABI::Windows::Foundation::AsyncStatus;
+using abi_IBuffer = ABI::Windows::Storage::Streams::IBuffer;
+using abi_IBufferByteAccess = ::Windows::Storage::Streams::IBufferByteAccess;
+
+// The M2AsyncWait function uses the non-blocking way to try to wait 
 // asynchronous call.
 //
 // Parameters:
 //
-// async
+// Async
 //     The asynchronous call you want to wait.
-// timeout
+// Timeout
 //     The maximum time interval for waiting the asynchronous call, in 
 //     milliseconds. A value of -1 indicates that the suspension should not 
 //     time out.
@@ -36,15 +44,15 @@ using Windows::Storage::Streams::IBufferByteAccess;
 //
 // If the function succeeds, the return value is S_OK.
 // If the function fails, the return value is the HRESULT error code.
-HRESULT m2_await_internal(ComPtr<IInspectable>& async, int32 timeout) throw()
+HRESULT M2AsyncWait(ComPtr<IInspectable>& Async, int32 Timeout) throw()
 {
 	HRESULT hr = S_OK;
 	HRESULT hrResult = S_OK;
 	ComPtr<IAsyncInfo> asyncInfo;
-	AsyncStatus status = AsyncStatus::Started;	
+	abi_AsyncStatus status = AsyncStatus::Started;	
 
 	// Get the IAsyncInfo interface.
-	hr = async.As(&asyncInfo);
+	hr = Async.As(&asyncInfo);
 	if (SUCCEEDED(hr))
 	{
 		// Wait the asynchronous call until the status is not Started or the 
@@ -52,15 +60,15 @@ HRESULT m2_await_internal(ComPtr<IInspectable>& async, int32 timeout) throw()
 		while (
 			SUCCEEDED(hr = asyncInfo->get_Status(&status)) &&
 			AsyncStatus::Started == status &&
-			(timeout == -1 || timeout > 0))
+			(Timeout == -1 || Timeout > 0))
 		{
 			// Calling SleepEx() for sleep 50ms every time because 
 			// Microsoft says that all UWP APIs that can't guarantee to 
 			// complete within 50ms has been made asynchronous and its name
 			// suffixed with Async.
 			::SleepEx(50, FALSE);
-			if (timeout != -1)
-				timeout -= 50;
+			if (Timeout != -1)
+				Timeout -= 50;
 		}
 
 		// Check the status if no error when call the interface.
@@ -94,8 +102,8 @@ HRESULT m2_await_internal(ComPtr<IInspectable>& async, int32 timeout) throw()
 	return (SUCCEEDED(hr) ? hrResult : hr);
 }
 
-// The M2GetPointerFromIBuffer function retrieves the raw pointer from the 
-// provided IBuffer object. 
+// The M2GetPointer function retrieves the raw pointer from the provided 
+// IBuffer object. 
 //
 // Parameters:
 //
@@ -111,14 +119,208 @@ HRESULT m2_await_internal(ComPtr<IInspectable>& async, int32 timeout) throw()
 // The lifetime of the returned buffer is controlled by the lifetime of the 
 // buffer object that's passed to this method. When the buffer has been 
 // released, the pointer becomes invalid and must not be used.
-byte* M2GetPointerFromIBuffer(IBuffer^ Buffer) throw()
+byte* M2GetPointer(IBuffer^ Buffer) throw()
 {
 	byte* pBuffer = nullptr;
-	ComPtr<IBufferByteAccess> bufferByteAccess;
+	ComPtr<abi_IBufferByteAccess> bufferByteAccess;
 	if (SUCCEEDED(M2GetInspectable(Buffer).As(&bufferByteAccess)))
 	{
 		bufferByteAccess->Buffer(&pBuffer);
 	}
 
 	return pBuffer;
+}
+
+class BufferReference : public RuntimeClass<
+	RuntimeClassFlags<RuntimeClassType::WinRtClassicComMix>,
+	abi_IBuffer,
+	abi_IBufferByteAccess>
+{
+private:
+	UINT32 m_Capacity;
+	UINT32 m_Length;
+	byte* m_Pointer;
+
+public:
+	virtual ~BufferReference() throw()
+	{
+	}
+
+	STDMETHODIMP RuntimeClassInitialize(
+		byte* Pointer, UINT32 Capacity) throw()
+	{
+		m_Capacity = Capacity;
+		m_Length = Capacity;
+		m_Pointer = Pointer;
+		return S_OK;
+	}
+
+	// IBufferByteAccess::Buffer
+	STDMETHODIMP Buffer(byte **value) throw()
+	{
+		*value = m_Pointer;
+		return S_OK;
+	}
+
+	// IBuffer::get_Capacity
+	STDMETHODIMP get_Capacity(UINT32 *value) throw()
+	{
+		*value = m_Capacity;
+		return S_OK;
+	}
+
+	// IBuffer::get_Length
+	STDMETHODIMP get_Length(UINT32 *value) throw()
+	{
+		*value = m_Length;
+		return S_OK;
+	}
+
+	// IBuffer::put_Length
+	STDMETHODIMP put_Length(UINT32 value) throw()
+	{
+		if (value > m_Capacity)
+			return E_INVALIDARG;
+		m_Length = value;
+		return S_OK;
+	}
+};
+
+// The M2MakeIBuffer function retrieves the IBuffer object from the provided 
+// raw pointer.
+//
+// Parameters:
+//
+// Pointer
+//     The raw pointer you want to retrieve the IBuffer object.
+// Capacity
+//     The size of raw pointer you want to retrieve the IBuffer object.
+//
+// Return value:
+//
+// If the function succeeds, the return value is the IBuffer object from the 
+// provided raw pointer. If the function fails, the return value is nullptr.
+//
+// Warning: 
+// The lifetime of the returned IBuffer object is controlled by the lifetime of
+// the raw pointer that's passed to this method. When the raw pointer has been 
+// released, the IBuffer object becomes invalid and must not be used.
+IBuffer^ M2MakeIBuffer(byte* Pointer, UINT32 Capacity) throw()
+{
+	IBuffer^ buffer = nullptr;
+	
+	ComPtr<BufferReference> bufferReference;
+	if (SUCCEEDED(MakeAndInitialize<BufferReference>(
+		&bufferReference, Pointer, Capacity)))
+	{
+		buffer = reinterpret_cast<IBuffer^>(bufferReference.Get());
+	}
+
+	return buffer;
+}
+
+// The M2MakeUTF16String function converts from UTF-8 string to UTF-16 string.
+//
+// Parameters:
+//
+// UTF8String
+//     The UTF-8 string you want to convert.
+//
+// Return value:
+//
+// The return value is the UTF-16 string.
+wstring M2MakeUTF16String(const string& UTF8String)
+{
+	wstring UTF16String;
+
+	int UTF16StringLength = ::MultiByteToWideChar(
+		CP_UTF8,
+		0,
+		UTF8String.data(),
+		(int)UTF8String.size(),
+		nullptr,
+		0);
+	if (UTF16StringLength > 0)
+	{
+		UTF16String.resize(UTF16StringLength);
+		::MultiByteToWideChar(
+			CP_UTF8,
+			0,
+			UTF8String.data(),
+			(int)UTF8String.size(),
+			&UTF16String[0],
+			UTF16StringLength);
+	}
+
+	return UTF16String;
+}
+
+// The M2MakeUTF8String function converts from UTF-16 string to UTF-8 string.
+//
+// Parameters:
+//
+// UTF16String
+//     The UTF-16 string you want to convert.
+//
+// Return value:
+//
+// The return value is the UTF-8 string.
+string M2MakeUTF8String(const wstring& UTF16String)
+{
+	string UTF8String;
+
+	int UTF8StringLength = ::WideCharToMultiByte(
+		CP_UTF8,
+		0,
+		UTF16String.data(),
+		(int)UTF16String.size(),
+		nullptr,
+		0,
+		nullptr,
+		nullptr);
+	if (UTF8StringLength > 0)
+	{
+		UTF8String.resize(UTF8StringLength);
+		::WideCharToMultiByte(
+			CP_UTF8,
+			0,
+			UTF16String.data(),
+			(int)UTF16String.size(),
+			&UTF8String[0],
+			UTF8StringLength,
+			nullptr,
+			nullptr);
+	}
+
+	return UTF8String;
+}
+
+// The M2MakeUTF16String function converts from C++/CX string to UTF-16 string.
+//
+// Parameters:
+//
+// PlatformString
+//     The C++/CX string you want to convert.
+//
+// Return value:
+//
+// The return value is the UTF-16 string.
+wstring M2MakeUTF16String(String^& PlatformString)
+{
+	return wstring(PlatformString->Data(), PlatformString->Length());
+}
+
+// The M2MakeUTF8String function converts from C++/CX string to UTF-8 string.
+//
+// Parameters:
+//
+// PlatformString
+//     The C++/CX string you want to convert.
+//
+// Return value:
+//
+// The return value is the UTF-8 string.
+string M2MakeUTF8String(String^& PlatformString)
+{
+	return M2MakeUTF8String(M2MakeUTF16String(PlatformString));
 }